fix(cacheableRequest): prevent collisions on POST, PUT, and PATCH
If a server responds with valid cache-control headers POST, PUT, and PATCH
requests may be cached. This may lead to cache collisions if the URI is
the same between requests. In order to avoid this, this change will hash
the body of the request and add that to the cache key. One note: the cache
is completely bypassed in the event that the request body is a stream. This
is because the key could not be generated without waiting for the stream to
end and the contents of the stream may or may not themselves be easily
hashed.